Refaktorerte POM-struktur for bedre vedlikeholdbarhet
  Standardiserte assertion-navngivning og forenklet page object-strukturen
  ved å fjerne separat assertions-klasse og organisere koden i tydelige seksjoner.

  Endringer:
  - Standardiserte assert* → verifiser* (9 metoder i aarsavregning og send-brev)
  - Fjernet opprett-ny-sak-assertions.page.ts (242 linjer)
  - Flyttet alle assertions tilbake til OpprettNySakPage
  - Fjernet delegerings-metoder
  - Organisert metoder i seksjoner (ACTIONS, QUERIES, ASSERTIONS)
  - Oppdatert tester til nye metodenavn
